Rewrite constants and serverTime modules in TS
So that functions can be imported directly from modules (to not risk importing all pmcrypto functions) and preserve type checks
