--- lib/serverTime.js
@@ -1,9 +1,0 @@
-let lastServerTime = null;
-
-export const serverTime = () => lastServerTime || new Date();
-export const updateServerTime = (serverDate) => {
-    if (serverDate > lastServerTime) {
-        lastServerTime = serverDate;
-    }
-    return lastServerTime;
-};
